+++ src/Type/Payment/TransactionPartnerUser.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Vjik\TelegramBot\Api\Type\Payment;
 
+use Vjik\TelegramBot\Api\ParseResult\ValueProcessor\ArrayMap;
+use Vjik\TelegramBot\Api\ParseResult\ValueProcessor\PaidMediaValue;
 use Vjik\TelegramBot\Api\Type\User;
 
 /**
@@ -14,6 +16,8 @@ final readonly class TransactionPartnerUser implements TransactionPartner
     public function __construct(
         public User $user,
         public ?string $invoicePayload = null,
+        #[ArrayMap(PaidMediaValue::class)]
+        public ?array $paidMedia = null,
     ) {}
 
     public function getType(): string