+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Telegram Bot API for PHP Change Log
 
+## 0.4.2 under construction
+
+- New #117: Add `createChatSubscriptionInviteLink` and `editChatSubscriptionInviteLink` methods.
+- New #117: Add `businessConnectionId` parameter to `SendPaidMedia` method.
+- New #117: Add `ReactionTypePaid` type.
+- New #117: Add `paidMedia` field to `TransactionPartnerUser` type.
+- New #117: Add `subscriptionPeriod` and `subscriptionPrice` fields to `ChatInviteLink` type.
+- New #117: Add `untilDate` field to `ChatMemberMember` type.
+
 ## 0.4.1 August 1, 2024
 
 - New #115: Add `hasMainWebApp` field to `User` type.